frontend/requirements.txt
streamlit
pillow
requests
streamlit-webrtc
opencv-python-headless
numpy